+++ httpcore/_sync/http2.py
@@ -158,8 +158,8 @@ class SyncHTTP2Connection(SyncBaseHTTPConnection):
     def is_closed(self) -> bool:
         return False
 
-    def is_connection_dropped(self) -> bool:
-        return self.socket.is_connection_dropped()
+    def is_socket_readable(self) -> bool:
+        return self.socket.is_readable()
 
     def close(self) -> None:
         logger.trace("close_connection=%r", self)